requirements.txt
requests
asgiref==3.2.5
Django==3.0.4
pytz==2019.3
sqlparse==0.3.1
plaid-python==3.7.0
djangorestframework==3.11.0
django-environ
djangorestframework-camel-case
djoser==2.0.3
